requirements.txt
appdirs==1.4.3
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1.1
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
South==1.0.2
